front-end/requirements.txt
ijson>=3.2,<4
psycopg2-binary>=2.9,<3
requests>=2.31,<3